refactoring helpers to avoid double requires
