+++ test/api/directory_listings_controller_test.rb
@@ -1,5 +1,4 @@
 require 'api_test_helper'
-require 'integration_test_helper'
 
 class DirectoryListingsControllerTest < ActionDispatch::IntegrationTest
   include Devise::Test::IntegrationHelpers
