+++ test/api/study_shares_controller_test.rb
@@ -1,5 +1,4 @@
 require 'api_test_helper'
-require 'integration_test_helper'
 
 class StudySharesControllerTest < ActionDispatch::IntegrationTest
   include Devise::Test::IntegrationHelpers
