+++ test/api_test_helper.rb
@@ -1,4 +1,5 @@
 require 'simplecov_helper'
+require 'seeds_helper'
 
 ENV['RAILS_ENV'] ||= 'test'
 require_relative '../config/environment'
